+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/ProxySPITest.java
@@ -927,7 +927,7 @@ public class ProxySPITest extends AbstractPolyglotTest {
         }
     }
 
-    private void assertHostError(InteropCallable r) {
+    private static void assertHostError(InteropCallable r) throws InteropException {
         try {
             r.call();
             Assert.fail();
@@ -937,7 +937,7 @@ public class ProxySPITest extends AbstractPolyglotTest {
             InteropLibrary interop = InteropLibrary.getUncached();
             Assert.assertTrue(interop.isException(e));
             Assert.assertEquals("Host Error", e.getMessage());
-            Assert.assertTrue(languageEnv.asHostException(e) instanceof TestError);
+            Assert.assertTrue(INTEROP.asHostObject(e) instanceof TestError);
         }
     }