+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleStackTrace.java
@@ -47,7 +47,6 @@ import java.util.ListIterator;
 import java.util.Objects;
 
 import org.graalvm.polyglot.PolyglotException;
-import org.graalvm.polyglot.impl.AbstractPolyglotImpl.AbstractHostLanguageService;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.frame.Frame;
@@ -270,13 +269,7 @@ public final class TruffleStackTrace extends Exception {
     }
 
     private static boolean isHostException(Throwable throwable) {
-        Object polyglotEngine = LanguageAccessor.ENGINE.getCurrentPolyglotEngine();
-        if (polyglotEngine == null) {
-            return false;
-        }
-        AbstractHostLanguageService hostService = LanguageAccessor.ENGINE.getHostService(polyglotEngine);
-        // hostService is null during context pre-initialization
-        return hostService != null && hostService.isHostException(throwable);
+        return LanguageAccessor.ENGINE.isHostException(throwable);
     }
 
     private static final class TracebackElement {