+++ truffle/src/com.oracle.truffle.host/src/com/oracle/truffle/host/HostLanguageService.java
@@ -162,30 +162,8 @@ public class HostLanguageService extends AbstractHostLanguageService {
     }
 
     @Override
-    public boolean isHostValue(Object value) {
-        Object obj = HostLanguage.unwrapIfScoped(language, value);
-        return (obj instanceof HostObject) ||
-                        (obj instanceof HostFunction) ||
-                        (obj instanceof HostException) ||
-                        (obj instanceof HostProxy);
-    }
-
-    @Override
-    public Object unboxHostObject(Object hostValue) {
-        return HostObject.valueOf(language, hostValue);
-    }
-
-    @Override
     public Object unboxProxyObject(Object hostValue) {
         return HostProxy.toProxyHostObject(language, hostValue);
-    }
-
-    @Override
-    public Throwable unboxHostException(Throwable hostValue) {
-        if (hostValue instanceof HostException) {
-            return ((HostException) hostValue).getOriginal();
-        }
-        return null;
     }
 
     @Override
@@ -200,32 +178,8 @@ public class HostLanguageService extends AbstractHostLanguageService {
     }
 
     @Override
-    public boolean isHostException(Object exception) {
-        return exception instanceof HostException;
-    }
-
-    @Override
-    public boolean isHostFunction(Object value) {
-        return HostFunction.isInstance(language, value);
-    }
-
-    @Override
-    public boolean isHostObject(Object value) {
-        return HostObject.isInstance(language, value);
-    }
-
-    @Override
     public boolean isHostProxy(Object value) {
         return HostProxy.isProxyGuestObject(language, value);
-    }
-
-    @Override
-    public boolean isHostSymbol(Object obj) {
-        Object o = HostLanguage.unwrapIfScoped(language, obj);
-        if (o instanceof HostObject) {
-            return ((HostObject) o).isStaticClass();
-        }
-        return false;
     }
 
     @Override
@@ -261,7 +215,6 @@ public class HostLanguageService extends AbstractHostLanguageService {
     public Object migrateValue(Object targetContext, Object value, Object valueContext) {
         assert targetContext != valueContext;
         if (value instanceof TruffleObject) {
-            assert value instanceof TruffleObject;
             if (HostObject.isInstance(language, value)) {
                 return HostObject.withContext(language, value, (HostContext) HostAccessor.ENGINE.getHostContext(targetContext));
             } else if (value instanceof HostProxy) {
@@ -271,7 +224,6 @@ public class HostLanguageService extends AbstractHostLanguageService {
                  * The only way this can happen is with Value.asValue(TruffleObject). If it happens
                  * otherwise, its wrong.
                  */
-                assert value instanceof TruffleObject;
                 return value;
             } else {
                 // cannot migrate
@@ -281,15 +233,6 @@ public class HostLanguageService extends AbstractHostLanguageService {
             assert InteropLibrary.isValidValue(value);
             return value;
         }
-    }
-
-    @Override
-    public Error toHostResourceError(Throwable hostException) {
-        Throwable t = unboxHostException(hostException);
-        if (t instanceof StackOverflowError || t instanceof OutOfMemoryError) {
-            return (Error) t;
-        }
-        return null;
     }
 
     @Override